Renamed directory and files
