+++ admin/public/favicon.ico
