+++ admin/public/manifest.json
