+++ admin/src/domains.js
@@ -5,8 +5,8 @@ export const PostList = (props) => (
     <List {...props}>
         <Datagrid>
             <TextField source="id" />
-            <TextField source="title" />
-            <TextField source="body" />
+            <TextField source="name" />
+            <TextField source="description" />
         </Datagrid>
     </List>
 );
