--- demo-admin/src/App.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { jsonServerRestClient, Admin, Resource } from 'admin-on-rest';
-
-import { PostList } from './posts';
-import authClient from './authClient';
-
-const App = () => (
-    <Admin authClient={authClient}
-           restClient={jsonServerRestClient('http://jsonplaceholder.typicode.com')}>
-        <Resource name="posts" list={PostList} />
-    </Admin>
-);
-
-export default App;
